+++ bindings/nodejs/lib/index.ts
@@ -5,3 +5,4 @@ export * from './client';
 export * from './wallet';
 export * from './secretManager';
 export * from './utils';
+export * from './types';