+++ bindings/nodejs/types/client/bridge/index.ts
@@ -43,7 +43,7 @@ import type {
     __NftOutputIdMethod__,
     __FoundryOutputIdsMethod__,
     __FoundryOutputIdMethod__,
-    __TryGetOutputsMethod__,
+    __GetOutputsIgnoreErrorsMethod__,
     __FindBlocksMethod__,
     __RetryMethod__,
     __RetryUntilIncludedMethod__,
@@ -116,7 +116,7 @@ export type __ClientMethods__ =
     | __NftOutputIdMethod__
     | __FoundryOutputIdsMethod__
     | __FoundryOutputIdMethod__
-    | __TryGetOutputsMethod__
+    | __GetOutputsIgnoreErrorsMethod__
     | __FindBlocksMethod__
     | __RetryMethod__
     | __RetryUntilIncludedMethod__