+++ bindings/python/examples/wallet/10-claim-outputs.py
@@ -1,15 +1,20 @@
 from iota_sdk import Wallet
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
 
 # In this example we will claim outputs that have additional unlock conditions as expiration or storage deposit return
-
-# Explorer url
-EXPLORER = "https://explorer.shimmer.network/testnet"
 
 wallet = Wallet('./alice-database')
 
 account = wallet.get_account('Alice')
 
-wallet.set_stronghold_password("some_hopefully_secure_password")
+if 'STRONGHOLD_PASSWORD' not in os.environ:
+    print(".env STRONGHOLD_PASSWORD is undefined, see .env.example")
+    sys.exit(1)
+
+wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 
 # Sync account with the node
 response = account.sync()
@@ -20,6 +25,5 @@ output_ids = account.get_outputs_with_additional_unlock_conditions('All')
 print(f'Available outputs to claim: {output_ids}')
 
 transaction = account.claim_outputs(output_ids)
+print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction["blockId"]}')
 
-print(f'Transaction: {transaction["transactionId"]}')
-print(f'Block sent: {EXPLORER}/block/" + {transaction["blockId"]}');