+++ sdk/CHANGELOG.md
@@ -32,6 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `OutputData::metadata` changed from `OutputMetadataDto` to `OutputMetadata`;
+- `Client::get_outputs()` and derived methods return `OutputWithMetadata` instead of `OutputWithMetadataDto`;
+- rename `Client::try_get_outputs()` into `Client::get_outputs_ignore_errors()`;
+- rename `Client::try_get_outputs_metadata()` into `Client::get_outputs_metadata_ignore_errors()`;
+- MQTT connections to a node using https will now use wss/tls with native certificates;
 
 ### Changed